MSFT: 50731161 - Enable level 4 warnings (#322)
## Why is this change being made?
I've been bitten twice by build breaks due to [C4018 (signed/unsigned mismatch)](https://learn.microsoft.com/en-us/cpp/error-messages/compiler-warnings/compiler-warning-level-3-c4018?view=msvc-170) warnings that repro'd for x86 but not x64 (#308 and #316). It turns out that for the x64 build the warnings being generated were [C4388 (signed/unsigned mismatch)](https://learn.microsoft.com/en-us/cpp/error-messages/compiler-warnings/c4388?view=msvc-170) which is a level 4, off by default warning.

We want to follow [best practices](https://learn.microsoft.com/en-us/cpp/code-quality/build-reliable-secure-programs?view=msvc-170#23-code-based-or-static-analysis) and enable level 4 warnings.

## What changed?
- Enabled level 4 warnings (`/W4`)
- Enabled [C4388 (signed/unsigned mismatch)](https://learn.microsoft.com/en-us/cpp/error-messages/compiler-warnings/c4388?view=msvc-170), which is a level 4 warning but off by default (`/w44388`)
- Addressed build breaks due to newly enabled warnings being treated as errors

## How was the change tested?
I validated the following scenarios:
- Ogg playback in MediaPlayerCPP/MediaPlayerCS
- Ogg playback in Media Player with WME (in-proc and out-of-proc)
- MKV with Vorbis/Theora playback in Media Player with WME (in-proc and out-of-proc)

+++ FFmpegInterop/BitstreamReader.cpp
@@ -50,7 +50,7 @@ namespace winrt::FFmpegInterop::implementation
 		auto [numBytesToSkip, numBitsToSkip] = div(numBits, BITS_PER_BYTE);
 
 		m_byteIndex += numBytesToSkip;
-		m_bitIndex += numBitsToSkip;
+		m_bitIndex += static_cast<uint8_t>(numBitsToSkip);
 		if (m_bitIndex >= BITS_PER_BYTE)
 		{
 			m_byteIndex++;
